+++ MIGRATION.md
@@ -1,0 +1,5 @@
+# Migration Guide
+
+## 1.3.1 -> 1.4.0
+
+TBD
