+++ tests/perpetual/limit_order_object/test_limit_order_object_settlement.py
@@ -15,7 +15,7 @@ async def test_create_buy_limit_order_settlement_data(
 ):
     mocker.patch("x10.utils.nonce.generate_nonce", return_value=FROZEN_NONCE)
 
-    from x10.configuration import MAINNET_CONFIG
+    from x10.perpetual.configuration import MAINNET_CONFIG
     from x10.perpetual.limit_order_object_settlement import create_order_settlement_data
 
     trading_account = create_trading_account()
