+++ tests/perpetual/order_object/test_conditional_order_object.py
@@ -5,7 +5,6 @@ from freezegun import freeze_time
 from hamcrest import assert_that, equal_to
 from pytest_mock import MockerFixture
 
-from x10.configuration import TESTNET_CONFIG
 from x10.models.order import (
     OrderPriceType,
     OrderSide,
@@ -13,6 +12,7 @@ from x10.models.order import (
     OrderTriggerPriceType,
     OrderType,
 )
+from x10.perpetual.configuration import TESTNET_CONFIG
 
 FROZEN_NONCE = 1473459052
