+++ tests/perpetual/order_object/test_tpsl_order_object.py
@@ -6,7 +6,6 @@ from freezegun import freeze_time
 from hamcrest import assert_that, equal_to
 from pytest_mock import MockerFixture
 
-from x10.configuration import TESTNET_CONFIG
 from x10.models.order import (
     OrderPriceType,
     OrderSide,
@@ -15,6 +14,7 @@ from x10.models.order import (
     OrderType,
     SelfTradeProtectionLevel,
 )
+from x10.perpetual.configuration import TESTNET_CONFIG
 from x10.utils.date import utc_now
 
 FROZEN_NONCE = 1473459052
