+++ tests/perpetual/test_transfer_object.py
@@ -5,7 +5,7 @@ from freezegun import freeze_time
 from hamcrest import assert_that, equal_to
 from pytest_mock import MockerFixture
 
-from x10.configuration import TESTNET_CONFIG
+from x10.perpetual.configuration import TESTNET_CONFIG
 
 FROZEN_NONCE = 1473459052
