+++ x10/core/stark_account.py
@@ -7,6 +7,11 @@ from x10.utils.string import is_hex_string
 
 
 class StarkPerpetualAccount:
+    """
+    Attributes:
+        __trading_fee (dict): Field is deprecated and will be removed.
+    """
+
     __vault: int
     __private_key: int
     __public_key: int
