+++ x10/perpetual/order_object.py
@@ -4,7 +4,6 @@ from decimal import Decimal
 from typing import Callable, Optional, Tuple
 
 from x10.config import DEFAULT_FEES
-from x10.configuration import StarknetDomain
 from x10.core.stark_account import StarkPerpetualAccount
 from x10.models.fee import TradingFeeModel
 from x10.models.market import MarketModel
@@ -21,6 +20,7 @@ from x10.models.order import (
     SelfTradeProtectionLevel,
     TimeInForce,
 )
+from x10.perpetual.configuration import StarknetDomain
 from x10.perpetual.order_object_settlement import (
     SettlementDataCtx,
     create_order_settlement_data,
