+++ x10/perpetual/user_client/user_client.py
@@ -7,9 +7,9 @@ from eth_account import Account
 from eth_account.messages import encode_defunct
 from eth_account.signers.local import LocalAccount
 
-from x10.configuration import EndpointConfig
 from x10.errors import X10Error
 from x10.models.account import AccountModel, ApiKeyRequestModel, ApiKeyResponseModel
+from x10.perpetual.configuration import EndpointConfig
 from x10.perpetual.user_client.onboarding import (
     OnboardedClientModel,
     StarkKeyPair,
